6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@ int main(void)
 {
 	long long int a;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of aint: %lu byte(s)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", sizeof(char));
+	printf("Size of aint: %lu byte(s)\n", sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(a));
+	printf("Size of a float: %lu byte(s)\n", sizeof(float));
 	return (0);
 }
